pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
#readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=6.1.0,<7.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "ruff (>=0.11.5,<0.12.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.config.settings"
pythonpath = [".", "app"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--reuse-db --nomigrations --cov=app --cov-report=term-missing"
testpaths = ["tests"]



